PowerUps Fix
- Deleted Swapping obstacles and collectables powerup (might add it back if we have time)
- Fixed the PowerUp that makes the player look like an obstacle

+++ Player.cs
@@ -102,7 +102,7 @@ namespace _VP_Project___Crazy_Eater
         }
         public void ResetRatio()
         {
-            Ratio = image.Width / image.Height;
+            Ratio = (float)image.Width / image.Height;
         }
     }
 }
